onging

+++ database/migrations/2021_08_13_120424_create_category_product_table.php
@@ -15,8 +15,8 @@ class CreateCategoryProductTable extends Migration
     {
         Schema::create('category_product', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('category_id');
-            $table->unsignedBigInteger('product_id');
+            $table->integer('category_id');
+            $table->integer('product_id');
             $table->timestamps();
         });
     }
